feat: Add delay directive
Moving the async display behavior into an delay directive. This could probably be done a lot cleaner with tests and code to handle input updates and early-destruction (i.e. destruction before the content is displayed). But it's mostly a demo of how little the tests need to change to handle such a refactoring. The changes to the tests are minimal and might arguably be changes that should have been there all along.